pyproject.toml
[tool.poetry]
name = "anime-chatbot"
version = "0.1.0"
description = ""
authors = ["Ritish017 <kurmaritish017@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.43.2"
langchain-groq = "^0.3.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
